feat(Added the ability to the menus to use the old links in the shanta_forager database): Added the ability to the menus to use the old links in the shanta_forager database
we attempted to get the code to load the links in the database to put
them in the session but we have not got the tt to actually list them. As
a work around till we fix this problem we will be using the old ttml
system to populate the lists.This work for the menus now. The ToDo
functions are not yet working.